src/requirements.txt
snowflake-connector-python=2.7.12
snowflake-snowpark-python=1.3.0
flask=2.0.3
